Move layers into part files
