+++ packages/flame_tiled/lib/src/renderable_tile_map.dart
@@ -10,6 +10,12 @@ import 'package:flame_tiled/src/simple_flips.dart';
 import 'package:flutter/painting.dart';
 import 'package:tiled/tiled.dart';
 
+part 'renderable_layers/renderable_layer.dart';
+part 'renderable_layers/tile_layer.dart';
+part 'renderable_layers/image_layer.dart';
+part 'renderable_layers/group_layer.dart';
+part 'renderable_layers/object_layer.dart';
+
 /// {@template _renderable_tiled_map}
 /// This is a wrapper over Tiled's [TiledMap] which can be rendered to a canvas.
 ///
@@ -232,637 +238,6 @@ class RenderableTiledMap {
     } on ArgumentError {
       return null;
     }
-  }
-}
-
-abstract class _RenderableLayer<T extends Layer> {
-  final T layer;
-
-  /// The parent [Group] layer (if it exists)
-  final _RenderableGroupLayer? parent;
-
-  _RenderableLayer(this.layer, this.parent);
-
-  bool get visible => layer.visible;
-
-  void render(Canvas canvas, Camera? camera);
-
-  void handleResize(Vector2 canvasSize) {}
-
-  void refreshCache() {}
-
-  double get offsetX {
-    return layer.offsetX + (parent?.offsetX ?? 0);
-  }
-
-  double get offsetY {
-    return layer.offsetY + (parent?.offsetY ?? 0);
-  }
-
-  double get opacity {
-    if (parent != null) {
-      return parent!.opacity * layer.opacity;
-    } else {
-      return layer.opacity;
-    }
-  }
-
-  double get parallaxX {
-    if (parent != null) {
-      return parent!.parallaxX * layer.parallaxX;
-    } else {
-      return layer.parallaxX;
-    }
-  }
-
-  double get parallaxY {
-    if (parent != null) {
-      return parent!.parallaxY * layer.parallaxY;
-    } else {
-      return layer.parallaxY;
-    }
-  }
-
-  /// Calculates the offset we need to apply to the canvas to compensate for
-  /// parallax positioning and scroll for the layer and the current camera
-  /// position
-  /// https://doc.mapeditor.org/en/latest/manual/layers/#parallax-scrolling-factor
-  void _applyParallaxOffset(Canvas canvas, Camera camera) {
-    final cameraX = camera.position.x;
-    final cameraY = camera.position.y;
-    final vpCenterX = camera.viewport.effectiveSize.x / 2;
-    final vpCenterY = camera.viewport.effectiveSize.y / 2;
-
-    // Due to how Tiled treats the center of the view as the reference
-    // point for parallax positioning (see Tiled docs), we need to offset the
-    // entire layer
-    var x = (1 - parallaxX) * vpCenterX;
-    var y = (1 - parallaxY) * vpCenterY;
-    // compensate the offset for zoom
-    x /= camera.zoom;
-    y /= camera.zoom;
-
-    // Now add the scroll for the current camera position
-    x += cameraX - (cameraX * parallaxX);
-    y += cameraY - (cameraY * parallaxY);
-
-    canvas.translate(x, y);
-  }
-}
-
-class _RenderableTileLayer extends _RenderableLayer<TileLayer> {
-  final TiledMap _map;
-  final Vector2 _destTileSize;
-  late final _layerPaint = ui.Paint();
-  late final Map<String, SpriteBatch> _cachedSpriteBatches;
-
-  _RenderableTileLayer(
-    super.layer,
-    super.parent,
-    this._map,
-    this._destTileSize,
-    this._cachedSpriteBatches,
-  ) {
-    _layerPaint.color = Color.fromRGBO(255, 255, 255, opacity);
-  }
-
-  @override
-  void refreshCache() {
-    _cacheLayerTiles();
-  }
-
-  void _cacheLayerTiles() {
-    for (final batch in _cachedSpriteBatches.values) {
-      batch.clear();
-    }
-
-    if (_map.orientation == null) {
-      return;
-    }
-
-    switch (_map.orientation!) {
-      case MapOrientation.isometric:
-        _cacheIsometricTiles();
-        break;
-      case MapOrientation.staggered:
-        _cacheIsometricStaggeredTiles();
-        break;
-      case MapOrientation.hexagonal:
-        _cacheHexagonalTiles();
-        break;
-      case MapOrientation.orthogonal:
-        _cacheOrthogonalLayerTiles();
-        break;
-    }
-  }
-
-  void _cacheOrthogonalLayerTiles() {
-    final tileData = layer.tileData!;
-    final batchMap = _cachedSpriteBatches;
-    final size = _destTileSize;
-
-    for (var ty = 0; ty < tileData.length; ty++) {
-      final tileRow = tileData[ty];
-
-      for (var tx = 0; tx < tileRow.length; tx++) {
-        final tileGid = tileRow[tx];
-        if (tileGid.tile == 0) {
-          continue;
-        }
-
-        final tile = _map.tileByGid(tileGid.tile);
-        final tileset = _map.tilesetByTileGId(tileGid.tile);
-        final img = tile.image ?? tileset.image;
-
-        if (img == null) {
-          continue;
-        }
-
-        final batch = batchMap[img.source];
-        if (batch == null) {
-          continue;
-        }
-
-        final src = tileset.computeDrawRect(tile).toRect();
-        final flips = SimpleFlips.fromFlips(tileGid.flips);
-        final scale = size.x / src.width;
-        final anchorX = src.width / 2;
-        final anchorY = src.height / 2;
-
-        late double offsetX;
-        late double offsetY;
-        offsetX = (tx + .5) * size.x;
-        offsetY = (ty + .5) * size.y;
-
-        final scos = flips.cos * scale;
-        final ssin = flips.sin * scale;
-
-        batch.addTransform(
-          source: src,
-          transform: ui.RSTransform(
-            scos,
-            ssin,
-            offsetX + -scos * anchorX + ssin * anchorY,
-            offsetY + -ssin * anchorX - scos * anchorY,
-          ),
-          flip: flips.flip,
-        );
-      }
-    }
-  }
-
-  void _cacheIsometricTiles() {
-    final tileData = layer.tileData!;
-    final batchMap = _cachedSpriteBatches;
-    final halfDestinationTile = _destTileSize / 2;
-    final size = _destTileSize;
-    final isometricXShift = _map.width * size.x * 0.5;
-
-    for (var ty = 0; ty < tileData.length; ty++) {
-      final tileRow = tileData[ty];
-
-      for (var tx = 0; tx < tileRow.length; tx++) {
-        final tileGid = tileRow[tx];
-        if (tileGid.tile == 0) {
-          continue;
-        }
-
-        final tile = _map.tileByGid(tileGid.tile);
-        final tileset = _map.tilesetByTileGId(tileGid.tile);
-        final img = tile.image ?? tileset.image;
-        if (img == null) {
-          continue;
-        }
-
-        final batch = batchMap[img.source];
-        if (batch == null) {
-          continue;
-        }
-
-        final src = tileset.computeDrawRect(tile).toRect();
-        final flips = SimpleFlips.fromFlips(tileGid.flips);
-        final scale = size.x / src.width;
-        final anchorX = src.width / 2;
-        final anchorY = src.height / 2;
-
-        late double offsetX;
-        late double offsetY;
-
-        offsetX = halfDestinationTile.x * (tx - ty) + isometricXShift;
-        offsetY = halfDestinationTile.y * (tx + ty) - size.y;
-
-        final scos = flips.cos * scale;
-        final ssin = flips.sin * scale;
-
-        batch.addTransform(
-          source: src,
-          transform: ui.RSTransform(
-            scos,
-            ssin,
-            offsetX + -scos * anchorX + ssin * anchorY,
-            offsetY + -ssin * anchorX - scos * anchorY,
-          ),
-          flip: flips.flip,
-        );
-      }
-    }
-  }
-
-  void _cacheIsometricStaggeredTiles() {
-    final tileData = layer.tileData!;
-    final batchMap = _cachedSpriteBatches;
-    final halfDestinationTile = _destTileSize / 2;
-    final size = _destTileSize;
-    final halfMapTile = Vector2(_map.tileWidth / 2, _map.tileHeight / 2);
-
-    var staggerY = 0.0;
-    var staggerX = 0.0;
-    // Hexagonal Ponity Tiles move down by a fractional amount.
-    if (_map.staggerAxis == StaggerAxis.y) {
-      staggerY = size.y * 0.5;
-    } else
-    // Hexagonal Flat Tiles move right by a fractional amount.
-    if (_map.staggerAxis == StaggerAxis.x) {
-      staggerX = size.x * 0.5;
-    }
-
-    for (var ty = 0; ty < tileData.length; ty++) {
-      final tileRow = tileData[ty];
-
-      // Hexagonal Pointy Tiles shift left and right depending on the row
-      if (_map.staggerAxis == StaggerAxis.y) {
-        if ((ty.isOdd && _map.staggerIndex == StaggerIndex.odd) ||
-            (ty.isEven && _map.staggerIndex == StaggerIndex.even)) {
-          staggerX = halfDestinationTile.x;
-        } else {
-          staggerX = 0.0;
-        }
-      }
-
-      // When staggering in the X axis, we need to hold painting of "lower"
-      // tiles (those with staggerY adjustments) otherwise they'll just get
-      // painted over. See the second pass loop after tx.
-      final xSecondPass = <_Transform>[];
-
-      for (var tx = 0; tx < tileRow.length; tx++) {
-        final tileGid = tileRow[tx];
-        if (tileGid.tile == 0) {
-          continue;
-        }
-
-        final tile = _map.tileByGid(tileGid.tile);
-        final tileset = _map.tilesetByTileGId(tileGid.tile);
-        final img = tile.image ?? tileset.image;
-        if (img == null) {
-          continue;
-        }
-
-        final batch = batchMap[img.source];
-        if (batch == null) {
-          continue;
-        }
-
-        // Tiles shift up and down as we move across the row.
-        if (_map.staggerAxis == StaggerAxis.x) {
-          if ((tx.isOdd && _map.staggerIndex == StaggerIndex.odd) ||
-              (tx.isEven && _map.staggerIndex == StaggerIndex.even)) {
-            staggerY = halfDestinationTile.y;
-          } else {
-            staggerY = 0.0;
-          }
-        }
-
-        final src = tileset.computeDrawRect(tile).toRect();
-        final flips = SimpleFlips.fromFlips(tileGid.flips);
-        final scale = size.x / src.width;
-        final anchorX = src.width - halfMapTile.x;
-        final anchorY = src.height - halfMapTile.y;
-
-        late double offsetX;
-        late double offsetY;
-
-        // halfTile.x: shifts the map half a tile forward rather than
-        //             lining up on at the center.
-        // halfTile.y: shifts the map half a tile down rather than
-        //             lining up on at the center.
-        // StaggerX/Y: Moves the tile forward/down depending on orientation.
-        //  * stagger: Isometric tiles move down or right by only a fraction,
-        //             specifically 1/2 the width or height, for packing.
-        if (_map.staggerAxis == StaggerAxis.y) {
-          offsetX = tx * size.x + staggerX + halfDestinationTile.x;
-          offsetY = ty * staggerY + halfDestinationTile.y;
-        } else {
-          offsetX = tx * staggerX + halfDestinationTile.x;
-          offsetY = ty * size.y + staggerY + halfDestinationTile.y;
-        }
-
-        final scos = flips.cos * scale;
-        final ssin = flips.sin * scale;
-
-        final transform = ui.RSTransform(
-          scos,
-          ssin,
-          offsetX + -scos * anchorX + ssin * anchorY,
-          offsetY + -ssin * anchorX - scos * anchorY,
-        );
-
-        // A second pass is only needed in the case of staggery.
-        if (_map.staggerAxis == StaggerAxis.x && staggerY > 0) {
-          xSecondPass.add(_Transform(src, transform, flips.flip, batch));
-        } else {
-          batch.addTransform(
-            source: src,
-            transform: transform,
-            flip: flips.flip,
-          );
-        }
-      }
-
-      for (final tile in xSecondPass) {
-        tile.batch.addTransform(
-          source: tile.source,
-          transform: tile.transform,
-          flip: tile.flip,
-        );
-      }
-    }
-  }
-
-  void _cacheHexagonalTiles() {
-    final tileData = layer.tileData!;
-    final batchMap = _cachedSpriteBatches;
-    final halfDestinationTile = _destTileSize / 2;
-    final size = _destTileSize;
-    final halfMapTile = Vector2(_map.tileWidth / 2, _map.tileHeight / 2);
-
-    var staggerY = 0.0;
-    var staggerX = 0.0;
-    // Hexagonal Ponity Tiles move down by a fractional amount.
-    if (_map.staggerAxis == StaggerAxis.y) {
-      staggerY = size.y * 0.75;
-    } else
-    // Hexagonal Flat Tiles move right by a fractional amount.
-    if (_map.staggerAxis == StaggerAxis.x) {
-      staggerX = size.x * 0.75;
-    }
-
-    for (var ty = 0; ty < tileData.length; ty++) {
-      final tileRow = tileData[ty];
-
-      // Hexagonal Pointy Tiles shift left and right depending on the row
-      if (_map.staggerAxis == StaggerAxis.y) {
-        if ((ty.isOdd && _map.staggerIndex == StaggerIndex.odd) ||
-            (ty.isEven && _map.staggerIndex == StaggerIndex.even)) {
-          staggerX = halfDestinationTile.x;
-        } else {
-          staggerX = 0.0;
-        }
-      }
-
-      // When staggering in the X axis, we need to hold painting of "lower"
-      // tiles (those with staggerY adjustments) otherwise they'll just get
-      // painted over. See the second pass loop after tx.
-      final xSecondPass = <_Transform>[];
-
-      for (var tx = 0; tx < tileRow.length; tx++) {
-        final tileGid = tileRow[tx];
-        if (tileGid.tile == 0) {
-          continue;
-        }
-
-        final tile = _map.tileByGid(tileGid.tile);
-        final tileset = _map.tilesetByTileGId(tileGid.tile);
-        final img = tile.image ?? tileset.image;
-        if (img == null) {
-          continue;
-        }
-
-        final batch = batchMap[img.source];
-        if (batch == null) {
-          continue;
-        }
-
-        // Hexagonal Flat tiles shift up and down as we move across the row.
-        if (_map.staggerAxis == StaggerAxis.x) {
-          if ((tx.isOdd && _map.staggerIndex == StaggerIndex.odd) ||
-              (tx.isEven && _map.staggerIndex == StaggerIndex.even)) {
-            staggerY = halfDestinationTile.y;
-          } else {
-            staggerY = 0.0;
-          }
-        }
-
-        final src = tileset.computeDrawRect(tile).toRect();
-        final flips = SimpleFlips.fromFlips(tileGid.flips);
-        final scale = size.x / src.width;
-        final anchorX = src.width - halfMapTile.x;
-        final anchorY = src.height - halfMapTile.y;
-
-        late double offsetX;
-        late double offsetY;
-
-        // halfTile.x: shifts the map half a tile forward rather than
-        //             lining up on at the center.
-        // halfTile.y: shifts the map half a tile down rather than
-        //             lining up on at the center.
-        // StaggerX/Y: Moves the tile forward/down depending on orientation.
-        //  * stagger: Hexagonal tiles move down or right by only a fraction,
-        //             specifically 3/4 the width or height, for packing.
-        if (_map.staggerAxis == StaggerAxis.y) {
-          offsetX = tx * size.x + staggerX + halfDestinationTile.x;
-          offsetY = ty * staggerY + halfDestinationTile.y;
-        } else {
-          offsetX = tx * staggerX + halfDestinationTile.x;
-          offsetY = ty * size.y + staggerY + halfDestinationTile.y;
-        }
-
-        final scos = flips.cos * scale;
-        final ssin = flips.sin * scale;
-
-        final transform = ui.RSTransform(
-          scos,
-          ssin,
-          offsetX + -scos * anchorX + ssin * anchorY,
-          offsetY + -ssin * anchorX - scos * anchorY,
-        );
-
-        // A second pass is only needed in the case of staggery.
-        if (_map.staggerAxis == StaggerAxis.x && staggerY > 0) {
-          xSecondPass.add(_Transform(src, transform, flips.flip, batch));
-        } else {
-          batch.addTransform(
-            source: src,
-            transform: transform,
-            flip: flips.flip,
-          );
-        }
-      }
-
-      for (final tile in xSecondPass) {
-        tile.batch.addTransform(
-          source: tile.source,
-          transform: tile.transform,
-          flip: tile.flip,
-        );
-      }
-    }
-  }
-
-  @override
-  void render(Canvas canvas, Camera? camera) {
-    canvas.save();
-
-    canvas.translate(offsetX, offsetY);
-
-    if (camera != null) {
-      _applyParallaxOffset(canvas, camera);
-    }
-
-    for (final batch in _cachedSpriteBatches.values) {
-      batch.render(canvas, paint: _layerPaint);
-    }
-
-    canvas.restore();
-  }
-
-  static Future<_RenderableLayer> load(
-    TileLayer layer,
-    _RenderableGroupLayer? parent,
-    TiledMap map,
-    Vector2 destTileSize,
-  ) async {
-    return _RenderableTileLayer(
-      layer,
-      parent,
-      map,
-      destTileSize,
-      await _loadImages(map),
-    );
-  }
-
-  static Future<Map<String, SpriteBatch>> _loadImages(TiledMap map) async {
-    final result = <String, SpriteBatch>{};
-
-    for (final img in map.tiledImages()) {
-      final src = img.source;
-      if (src != null) {
-        result[src] = await SpriteBatch.load(src);
-      }
-    }
-
-    return result;
-  }
-}
-
-class _RenderableImageLayer extends _RenderableLayer<ImageLayer> {
-  final Image _image;
-  late final ImageRepeat _repeat;
-  Rect _paintArea = Rect.zero;
-
-  _RenderableImageLayer(super.layer, super.parent, this._image) {
-    _initImageRepeat();
-  }
-
-  @override
-  void handleResize(Vector2 canvasSize) {
-    _paintArea = Rect.fromLTWH(0, 0, canvasSize.x, canvasSize.y);
-  }
-
-  @override
-  void render(Canvas canvas, Camera? camera) {
-    canvas.save();
-
-    canvas.translate(offsetX, offsetY);
-
-    if (camera != null) {
-      _applyParallaxOffset(canvas, camera);
-    }
-
-    paintImage(
-      canvas: canvas,
-      rect: _paintArea,
-      image: _image,
-      opacity: opacity,
-      alignment: Alignment.topLeft,
-      repeat: _repeat,
-    );
-
-    canvas.restore();
-  }
-
-  void _initImageRepeat() {
-    if (layer.repeatX && layer.repeatY) {
-      _repeat = ImageRepeat.repeat;
-    } else if (layer.repeatX) {
-      _repeat = ImageRepeat.repeatX;
-    } else if (layer.repeatY) {
-      _repeat = ImageRepeat.repeatY;
-    } else {
-      _repeat = ImageRepeat.noRepeat;
-    }
-  }
-
-  static Future<_RenderableLayer> load(
-    ImageLayer layer,
-    _RenderableGroupLayer? parent,
-    Camera? camera,
-  ) async {
-    return _RenderableImageLayer(
-      layer,
-      parent,
-      await Flame.images.load(layer.image.source!),
-    );
-  }
-}
-
-class _RenderableGroupLayer extends _RenderableLayer<Group> {
-  /// The child layers of this [Group] to be rendered recursively.
-  ///
-  /// NOTE: This is set externally instead of via constructor params because
-  ///       there are cyclic dependencies when loading the renderable layers.
-  late final List<_RenderableLayer> children;
-
-  _RenderableGroupLayer(
-    super.layer,
-    super.parent,
-  );
-
-  @override
-  void refreshCache() {
-    for (final child in children) {
-      child.refreshCache();
-    }
-  }
-
-  @override
-  void handleResize(Vector2 canvasSize) {
-    for (final child in children) {
-      child.handleResize(canvasSize);
-    }
-  }
-
-  @override
-  void render(ui.Canvas canvas, Camera? camera) {
-    for (final child in children) {
-      child.render(canvas, camera);
-    }
-  }
-}
-
-class _UnrenderableLayer extends _RenderableLayer {
-  _UnrenderableLayer(super.layer, super.parent);
-
-  @override
-  void render(Canvas canvas, Camera? camera) {
-    // nothing to do
-  }
-
-  // ignore unrenderable layers when looping over the layers to render
-  @override
-  bool get visible => false;
-
-  static Future<_RenderableLayer> load(Layer layer) async {
-    return _UnrenderableLayer(layer, null);
   }
 }
